Start generalizing Open AI based bindings.
